requirements.txt
cryptography~=38.0.1
discord~=2.0.0
requests~=2.28.1
PyYAML~=6.0
psutil~=5.9.3
PyJWT~=2.6.0
pywin32~=304
ujson~=5.5.0
